Code refactoring for dialog components...
